+++ perf_tests/ze_peer/src/ze_peer_parallel_single_target.cpp
@@ -64,7 +64,7 @@ void ZePeer::perform_bidirectional_parallel_copy_to_single_target(
   Timer<std::chrono::microseconds::period> timer;
 
   /* Warm up */
-  for (int i = 0; i < warm_up_iterations; i++) {
+  for (uint32_t i = 0U; i < warm_up_iterations; i++) {
     SUCCESS_OR_TERMINATE(zeEventHostSignal(event));
     for (size_t e = 0; e < num_engines; e++) {
       SUCCESS_OR_TERMINATE(zeCommandQueueExecuteCommandLists(
@@ -88,7 +88,7 @@ void ZePeer::perform_bidirectional_parallel_copy_to_single_target(
 
   do {
     timer.start();
-    for (int i = 0; i < number_iterations; i++) {
+    for (uint32_t i = 0U; i < number_iterations; i++) {
       for (size_t e = 0; e < num_engines; e++) {
         SUCCESS_OR_TERMINATE(zeCommandQueueExecuteCommandLists(
             ze_peer_devices[local_device_id].engines[queues[e]].first, 1,
@@ -157,7 +157,7 @@ void ZePeer::perform_parallel_copy_to_single_target(
   Timer<std::chrono::microseconds::period> timer;
 
   /* Warm up */
-  for (int i = 0; i < warm_up_iterations; i++) {
+  for (uint32_t i = 0U; i < warm_up_iterations; i++) {
     for (size_t e = 0; e < num_engines; e++) {
       SUCCESS_OR_TERMINATE(zeCommandQueueExecuteCommandLists(
           ze_peer_devices[queue_device_id].engines[queues[e]].first, 1,
@@ -178,7 +178,7 @@ void ZePeer::perform_parallel_copy_to_single_target(
 
   do {
     long double time_usec = 0;
-    for (int i = 0; i < number_iterations; i++) {
+    for (uint32_t i = 0U; i < number_iterations; i++) {
       for (size_t e = 0; e < num_engines; e++) {
         SUCCESS_OR_TERMINATE(zeCommandQueueExecuteCommandLists(
             ze_peer_devices[queue_device_id].engines[queues[e]].first, 1,
@@ -211,7 +211,7 @@ void ZePeer::perform_parallel_copy_to_single_target(
 
 void ZePeer::bandwidth_latency_parallel_to_single_target(
     peer_test_t test_type, peer_transfer_t transfer_type,
-    int number_buffer_elements, uint32_t remote_device_id,
+    size_t number_buffer_elements, uint32_t remote_device_id,
     uint32_t local_device_id) {
 
   size_t buffer_size = 0;
